requirements.txt
PyYAML==5.3.1
black==19.10b0
dacite==1.4.0
isort==4.3.21
kubernetes==11.0.0
mypy==0.770
